rename ChipRadioButtonGroup
